+++ src/librustc_session/lib.rs
@@ -1,3 +1,4 @@
+#![feature(crate_visibility_modifier)]
 #![feature(test)]
 
 // Use the test crate here so we depend on getopts through it. This allow tools to link to both
@@ -13,7 +14,9 @@ pub mod node_id;
 pub mod parse;
 
 mod code_stats;
+#[macro_use]
 pub mod config;
+mod options;
 pub mod filesearch;
 pub mod search_paths;
